Automatically generate certs
This was a pain and took way too long!  We now use `kustomize` to
template in the generated certs.  However, to get the caBundle into the
webhook config we have to use sed because that does not seem to be
templateable yet.

Not fun, but at least now it works?  Might have been much easier to just
copy-paste the cert converted to base64 whenever it was necessary.  But
oh well, now it's automated.

(It might also have been easier to "just use Kuh" at this point.)