Prepare validating AccessRequest in webhook
This does not validate anything yet, but it sets up everything so we can
validate in the future.  This was quite the drag, because for Kubernetes
to parse objects properly they need to implement deepcopy, which should
be code-generated, which requires a GOPATH, which... was a bit of a
nightmare to set up.

But now it is set up and we _should_ be able to validate access requests
very soon!