Make access request owner of associated objects
This makes it easy to then remove everything again upon use, which
Kubernetes now handles for us.

One potential downside: It seems to be possible to specify non-existing
objects as owner references.  I.e. before setting the APIVersion to
"spreadgroup.com/v1" I had it set to "v1", which would work but then not delete
the objects (likely because they could not be found).

We may want to delete objects in the foreground and/or enable
"BlockOwnerDeletion".