Delete unused and expired objects in the background
Previously objects would only be deleted if they were attempted to be
used.  Now they get cleaned up in the background.

They need to be handled separately because of the `validFor` field.
Without the field they will be deleted after a certain interval of not
being used.  With the field we know when they should expire and can
delete them after that.